PIC-firmware and CUI-Zero proyects added with documentation in doxygen. Deleted old version of PIC-firmware.
