+++ doc/groups.dox
@@ -37,6 +37,7 @@ namespace teo {}
  * @brief Contains the firmware that you need to program the Cui Absolute encoders.
  */
 
+
 /**
  * \defgroup teo_body_programs teo-body Programs
  *
