--- firmware/CuiAbsolute/PIC_source_2.0/CuiAbsolute.tagsrc
@@ -1,2 +1,0 @@
-C:\Users\Raul\Dropbox\RoboticsLab\Cui Absolute\Nuevos comentarios PIC_source\PIC_source_2.0\main.c
-C:\Users\Raul\Dropbox\RoboticsLab\Cui Absolute\Nuevos comentarios PIC_source\PIC_source_2.0\ECAN.h
