+++ firmware/CuiAbsolute/Pic_source/CuiAbsolute.mcp
