--- firmware/CuiAbsolute/Pic_source/ECAN.c.ini
@@ -1,7 +1,0 @@
-[Bookmarks]
-Count=0
-[Position]
-Line=16
-Column=28
-[FoldedLines]
-Count=0
