--- firmware/CuiAbsolute/Pic_source/EncoderRoman.mcp
@@ -1,62 +1,0 @@
-[HEADER]
-magic_cookie={66E99B07-E706-4689-9E80-9B2582898A13}
-file_version=1.0
-device=PIC18F2580
-[PATH_INFO]
-BuildDirPolicy=BuildDirIsProjectDir
-dir_src=
-dir_bin=
-dir_tmp=
-dir_sin=
-dir_inc=
-dir_lib=
-dir_lkr=
-[CAT_FILTERS]
-filter_src=*.asm;*.c
-filter_inc=*.h;*.inc
-filter_obj=*.o
-filter_lib=*.lib
-filter_lkr=*.lkr
-[CAT_SUBFOLDERS]
-subfolder_src=
-subfolder_inc=
-subfolder_obj=
-subfolder_lib=
-subfolder_lkr=
-[FILE_SUBFOLDERS]
-file_000=.
-file_001=.
-file_002=.
-file_003=.
-[GENERATED_FILES]
-file_000=no
-file_001=no
-file_002=no
-file_003=no
-[OTHER_FILES]
-file_000=no
-file_001=no
-file_002=yes
-file_003=yes
-[FILE_INFO]
-file_000=main.c
-file_001=ECAN.h
-file_002=ECAN.c
-file_003=ECAN.DEF
-[SUITE_INFO]
-suite_guid={5B7D72DD-9861-47BD-9F60-2BE967BF8416}
-suite_state=
-[TOOL_SETTINGS]
-TS{DD2213A8-6310-47B1-8376-9430CDFC013F}=
-TS{BFD27FBA-4A02-4C0E-A5E5-B812F3E7707C}=/o"$(BINDIR_)$(TARGETBASE).cof" /M"$(BINDIR_)$(TARGETBASE).map" /W
-TS{C2AF05E7-1416-4625-923D-E114DB6E2B96}=-Ou- -Ot- -Ob- -Op- -Or- -Od- -Opa-
-TS{ADE93A55-C7C7-4D4D-A4BA-59305F7D0391}=
-[INSTRUMENTED_TRACE]
-enable=0
-transport=0
-format=0
-[CUSTOM_BUILD]
-Pre-Build=
-Pre-BuildEnabled=1
-Post-Build=
-Post-BuildEnabled=1
