--- firmware/CuiAbsolute/Pic_source/Makefile
@@ -1,19 +1,0 @@
-# MPLAB IDE generated this makefile for use with GNU make.
-# Project: EncoderRoman.mcp
-# Date: Wed Aug 05 13:57:15 2015
-
-AS = MPASMWIN.exe
-CC = mcc18.exe
-LD = mplink.exe
-AR = mplib.exe
-RM = rm
-
-EncoderRoman.cof : main.o
-	$(LD) /p18F2580 "main.o" /u_CRUNTIME /z__MPLAB_BUILD=1 /o"EncoderRoman.cof" /M"EncoderRoman.map" /W
-
-main.o : main.c ECAN.h ECAN.c ../../../../Archivos\ de\ programa/Microchip/mplabc18/v3.47/h/delays.h ../../../../Archivos\ de\ programa/Microchip/mplabc18/v3.47/h/timers.h ../../../../Archivos\ de\ programa/Microchip/mplabc18/v3.47/h/spi.h ../../../../Archivos\ de\ programa/Microchip/mplabc18/v3.47/h/stdlib.h ../../../../Archivos\ de\ programa/Microchip/mplabc18/v3.47/h/stdio.h main.c ../../../../Archivos\ de\ programa/Microchip/mplabc18/v3.47/h/p18F2580.h ecan.def ../../../../Archivos\ de\ programa/Microchip/mplabc18/v3.47/h/p18cxxx.h ../../../../Archivos\ de\ programa/Microchip/mplabc18/v3.47/h/p18f2580.h ../../../../Archivos\ de\ programa/Microchip/mplabc18/v3.47/h/pconfig.h ../../../../Archivos\ de\ programa/Microchip/mplabc18/v3.47/h/stdarg.h ../../../../Archivos\ de\ programa/Microchip/mplabc18/v3.47/h/stddef.h
-	$(CC) -p=18F2580 "main.c" -fo="main.o" -Ou- -Ot- -Ob- -Op- -Or- -Od- -Opa-
-
-clean : 
-	$(RM) "main.o" "EncoderRoman.cof" "EncoderRoman.hex"
-
