+++ libraries/BodyYarp/CuiAbsolute/ICanBusSharerImpl.cpp
@@ -60,10 +60,10 @@ bool teo::CuiAbsolute::startContinuousPublishing(uint8_t delay)
     uint8_t msgData[8] = {0x01, 0x01, delay, 0x00, 0x00, 0x00, 0x00, 0x00};
     if( ! send(0 , 8, msgData) )   // -- primer campo "cob" lo dejamos a 0 (este campo resulta desconocido para nosotros)
     {
-        CD_ERROR("Could not send \"startContinuousPublishing\" to Cui Absolute Encoders.\n");
+        CD_ERROR("Could not send \"startContinuousPublishing\" to Cui Absolute Encoder.\n");
         return false;
     }
-    CD_SUCCESS("Send: \"startContinuousPublishing\" to Cui Absolute Encoders.\n");
+    CD_SUCCESS("Send: \"startContinuousPublishing\" to Cui Absolute Encoder.\n");
     return true;
 }
 
@@ -75,10 +75,10 @@ bool teo::CuiAbsolute::startPullPublishing()
     uint8_t msgData[8] = {0x01, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00}; // -- Comienza a publicar mensajes en modo pulling (modo 2) sin delay
     if( ! send(0, 8, msgData) )   // -- utilizaremos la funcion "send" por ser una funcion publica en vez de la funcion privada sendRaw
     {
-        CD_ERROR("Could not send \"startPullPublishing\" to Cui Absolute Encoders.\n");
+        CD_ERROR("Could not send \"startPullPublishing\" to Cui Absolute Encoder.\n");
         return false;
     }
-    CD_SUCCESS("Send: \"startPullPublishing\" to Cui Absolute Encoders. \n");
+    CD_SUCCESS("Send: \"startPullPublishing\" to Cui Absolute Encoder. \n");
     return true;
 }
 
@@ -90,14 +90,30 @@ bool teo::CuiAbsolute::stopPublishingMessages()
     uint8_t msgData[8] = {0x02, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00}; // -- Para de publicar mensajes
     if( ! send(0, 8, msgData) )
     {
-        CD_ERROR("Could not send \"stopPublishingMessages\" to Cui Absolute Encoders. %s\n");
+        CD_ERROR("Could not send \"stopPublishingMessages\" to Cui Absolute Encoder. %s\n");
         return false;
     }
-    CD_SUCCESS("Send: \"stopPublishingMessages\" to Cui Absolute Encoders. %s\n");
+    CD_SUCCESS("Send: \"stopPublishingMessages\" to Cui Absolute Encoder. %s\n");
     return true;
 }
 
-// ----------------------------------------------------------------------------
+// ------------------------------------------------------------------------------
+/*** Esta función actualmente no se utiliza debido a que implica cierta peligrosidad...
+ *
+ * bool teo::CuiAbsolute::setZeroPosition()
+ * {
+ *
+ *   uint8_t msgData[8] = {0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00}; // -- Homing
+ *   if( ! send(0, 8, msgData) )
+ *   {
+ *       CD_ERROR("Could not send \"setZeroPosition\" to Cui Absolute Encoder. %s\n");
+ *       return false;
+ *   }
+ *   CD_SUCCESS("Send: \"setZeroPosition\" to Cui Absolute Encoder. %s\n");
+ *   return true;
+ * }
+*/
+// ------------------------------------------------------------------------------
 
 bool teo::CuiAbsolute::interpretMessage( can_msg * message)
 {
